fix(cmd): run command not found issue while compiling
