+++ .gitignore
@@ -20,5 +20,5 @@
 go.work
 
 test_wallet
-
+/build
 teleq.sqlite
